% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{colfun0}
\alias{colfun0}
\title{default colonization function}
\usage{
colfun0(co, xt)
}
\arguments{
\item{co}{a numeric vector of length two (p, A), specifying the logit-transformed colonization probability when abundance is zero, and the log-transformed abundance observed immediately after a colonization event}

\item{xt}{a number or numeric vector of abundances at time t, before colonization has occurred}
}
\value{
a numeric, including number or numeric vector of length xt, with predicted abundances after colonization has occurred
}
\description{
Simulates colonization events - events occur as a binomial random process with probability ilogit(p), and populations are seeded with abundance exp(A).
}
\keyword{colonization}
