% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Microcosm experimental data}
\format{
A data frame with 271 rows and 4 variables:
\describe{
  \item{treatment}{Experimental treatment}
  \item{number}{Replicate number}
  \item{time}{Day of experiment}
  \item{Chlamydomonas.terricola}{Species abundance}
}
}
\source{
Burgmer & Hillebrand 2011, Oikos 120:922-933.
}
\usage{
dat
}
\description{
A dataset containing the abundances of Chlamydomonas terricola growing
in a multi-species community. Includes 17 time steps covering 463 days in two
treatments: LSA ('low' temperature, 'stable' oscillations, and 'absence' of predators)
and LVA ('low' temperature, 'variable' oscillations, and 'absence' of predators).
}
\keyword{datasets}
