% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{density_fun0}
\alias{density_fun0}
\title{Default density function for prior}
\usage{
density_fun0(param, pars = pars, priorsd = c(0.5, 0.5, 1, 0.5, 0.5,
  0.5))
}
\arguments{
\item{param}{a vector of length 6 or 8 with mean model parameters (or desired centers for prior distributions)}

\item{pars}{a list of parameter values, structured following the parseparam0 function, specifying the centers for the priors}

\item{priorsd}{a vector of length 6 or 8 with desired standard deviations for the prior distributions}
}
\value{
returns log likelihood of parameters given priors.
}
\description{
Default density function, following the syntax for priors in the BayesianTools package. Uses
lognormal priors for the intercept and slope of the observation error function, a normal prior for the
intercept of the process noise function, a lognormal for the slope of the process noise function,
a logit-normal prior for colonization probability, and a lognormal prior for the colonization abundance.
Note that prior likelihood is calculated based on values of param, which are centered around the listed values in variable pars.
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability,}
\keyword{time-series,}
