% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{EDMfun0}
\alias{EDMfun0}
\title{REDM deterministic function}
\source{
@source Adapted from Ye, Sugihara, et al. (2015), PNAS 112:E1569-E1576.
}
\usage{
EDMfun0(edmdat, xt, yblock, tm = NULL)
}
\arguments{
\item{edmdat}{A list of paramter values, which is passed on to the block_lnlp function. See block_lnlp for more details.}

\item{xt}{A matrix of dynamics for which abundances at time t+1 are to be predicted. Note that first column is assumed to represent t+1.}

\item{yblock}{A matrix of observed values, to be used as training data. Note that first column is assumed to represent t+1.}

\item{tm}{Time corresponding to position of xt in y, used to mask observations from training data. Defaults to NULL, meaning that no observations are masked.}
}
\value{
a number or numeric vector of length xt, with predicted abundances at time t+1
}
\description{
Estimates future states of xt based on based benavior
}
\keyword{EDM}
\keyword{applies}
\keyword{deterministic}
\keyword{dynamics}
\keyword{from}
\keyword{package}
\keyword{rEDM}
\keyword{reconstruct}
\keyword{series}
\keyword{the}
\keyword{time}
\keyword{to}
