% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{extend_particleFilter}
\alias{extend_particleFilter}
\title{Extend output from particle filter}
\source{
Adapted from Knape and Valpine (2012), Ecology 93:256-263.
}
\usage{
extend_particleFilter(pfout, pars, Next = 1000, detfun = detfun0,
  procfun = procfun0, obsfun = obsfun0, colfun = colfun0,
  edmdat = NULL, minval = 1e-16, trimq = c(0.01, 0.99),
  randstart = TRUE, concatts = FALSE)
}
\arguments{
\item{pfout}{Output from a previous run of the particleFilterLL function.}

\item{pars}{A list of parameter values. Must include elements obs (observation error parameters), proc (process noise parameters), and pcol (colonization parameters), which are passed on the their respecive functions, described below. If edmdat=NULL, then element det (deterministic process parameters) must be included.}

\item{Next}{Length of extended filter - i.e. number of timesteps to predict forward. Defaults to 1e3.}

\item{detfun}{A function that simulates deterministic dynamics, which takes in arguments sdet (parameters for deterministic model, taken from pars$proc), and xt, observed abundances at time t. Returns estimated abundances at time t+1 based on deterministic function (either a parametric function or an EDM function). Defaults to detfun0.}

\item{procfun}{A function that simulates process noise, which takes in arguments sp (parameters for process noise function, taken from pars$proc) and xt (abundances prior to process noise). Returns abundances after process noise has occurred. Defaults to procfun0.}

\item{obsfun}{An observation function, which takes in up to five variables, including so (a vector of parameter values, inherited from pars$obs), yt (a number, showing observed abundance at time t), xt (predicted abundances), binary value "inverse", and number "N". If inverse = TRUE,
then function should simulate N draws from the observation function, centered around value yt. If inverse = FALSE, then function should return log probability denisty of observed value yt given predicted values in xt. Defaults to obsfun0.}

\item{colfun}{A function simulating colonization events, that takes in two arguments: co, a vector of parameter values taken from pars$pcol, and xt, a number or numeric vector of abundances at time t, before colonization has occurred. Returns predicted abundances after colonization has occurred. Defaults to colful0.}

\item{edmdat}{A list including arguments to be passed to block_lnlp from rEDM package - see block_lnlp help file for details. Can also include optional matrix "extra_columns", a matrix with length(y) rows including extra covariates for attractor reconstruction, which defaults to NULL (i.e. no additional columns).
Default for edmdat is NULL, which implies that EDM will not be applied - instead, a detfun and pars$det must be included.}

\item{minval}{Minimum value, below will observations will be assumed to be zero. Can be helpful for simplex projections, as rounding errors can result in slightly nonzero values. Defaults to 1e-6.}

\item{trimq}{Either NULL (no limits), or a vector of length 2, including the lower and upper quantiles to keep for extinction and colonization projections. Can be helpful for excluding degenerate particles. Defaults to c(0.01, 0.99).}

\item{randstart}{Randomize starting positions within the time vector for the new particles? If FALSE, then all particles start at the end of the observed time series. Otherwise particles start at randomly chosen points, which can be helpful for getting better coverage of the full dynamic manifold. Defaults to TRUE.}

\item{concatts}{Concatenate original and extended particles? Defaults to FALSE.}
}
\value{
LL, P, rN, x, dem(col, mor)
}
\description{
Function for extending the particles (i.e. extrapolating timeseries into the future), based on output from particleFilterLL function.
Function is adapted from the R code of Knape and Valpine (2012), Ecology 93:256-263.
}
\keyword{Taylor}
\keyword{filter,}
\keyword{law}
\keyword{particle}
\keyword{power}
\keyword{stability,}
\keyword{time-series,}
