% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{getcm}
\alias{getcm}
\title{Get rates}
\usage{
getcm(dat)
}
\arguments{
\item{dat}{a numeric vector, including the timeseries}
}
\value{
a list including colonization and mortality probability per time step (pc and pm, respectively), and pocc, the expected fraction of time that the species will be present
}
\description{
Calculates colonization rate, mortality rate, and expected mean occupancy time based on a time series
}
\keyword{stability}
\keyword{time-series}
