% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_funs.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit}
\usage{
ilogit(x, ...)
}
\arguments{
\item{x}{a number, vector, matrix, etc. to be transformed from (-inf, inf) to (0 1) by the inverse logit transform}

\item{...}{additional arguments to be passed to plogis}
}
\value{
transformed result
}
\description{
Returns the inverse logit transformation of x
}
\keyword{logit}
