% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{indexsort}
\alias{indexsort}
\title{Sort output of particle filter}
\usage{
indexsort(fulltracemat, fulltraceindex, nsmp = NULL)
}
\arguments{
\item{fulltracemat}{full output of particles from the particleFilterLL function}

\item{fulltraceindex}{full output of particle indices from the particleFilterLL function}

\item{nsmp}{number of particle paths to sample - defaults to NULL, which samples all paths}
}
\value{
an index-sorted matrix - each column shows the trajectory of a single particle
}
\description{
Sorts outputs of particle filter based on index - returns a sorted list of particles, based on the
sampling trajectory through time. This is a somewhat more accurate estiamte of the true posterior than
are the stepwise samples provided by the filter.
}
\keyword{filter}
\keyword{particle}
