% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{inv_fun0}
\alias{inv_fun0}
\title{Default inverse transormation function}
\usage{
inv_fun0(x)
}
\arguments{
\item{x}{an nxm matrix with}
}
\value{
returns back-transformed values of parameters
}
\description{
Takes in a matrix, where each column represents a parameter. Returns parameters in untransformed space.
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability,}
\keyword{time-series,}
