% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{likelihood_EDM_piecewise}
\alias{likelihood_EDM_piecewise}
\title{calculate likelihood for piecewise data}
\usage{
likelihood_EDM_piecewise(
  param,
  y,
  libuse_y,
  smap_coefs,
  Euse,
  tuse,
  N,
  colpar = c(logit(1e-06), log(0.1))
)
}
\arguments{
\item{param}{parameters to be passed to likelihood0 function}

\item{y}{the time series to be analyzed}

\item{libuse_y}{a matrix with two columns, specifying the start end end positions of segments within vector y}

\item{smap_coefs}{a matrix of s-mapping coefficients}

\item{Euse}{embedding dimension for the s-mapping analysis}

\item{tuse}{theta for s-mapping analysis}

\item{N}{number of particles}

\item{colpar}{parameters to be passed to the colfun0 - defaults to c(logit(1e-6), log(0.1))}
}
\value{
summed log likelihood across all segments
}
\description{
Calculates likelihoods across several segments of data - e.g. multiple plots from a single experiment.
See documentation for particleFilterLL_piecewise for examples of use.
}
\keyword{dewdrop}
\keyword{filter}
\keyword{particle}
\keyword{regression}
