% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_funs.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(x, ...)
}
\arguments{
\item{x}{a number, vector, matrix, etc. to be transformed from (0, 1) to (-inf inf) by the logit transform}

\item{...}{additional arguments to be passed to plogis}
}
\value{
transformed result - impossible values are replaced with NA, without warnings
}
\description{
Returns the logit transformation of x
}
\keyword{logit}
