% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_funs.R
\name{logitnormal_imode}
\alias{logitnormal_imode}
\title{Get inverse logit-normal mode}
\usage{
logitnormal_imode(mu, sd)
}
\arguments{
\item{mu}{the value around which the mode should be centered (in logit space)}

\item{sd}{the standard deviation of the logit distribution (in logit space)}
}
\value{
the proposed mean for the distribution
}
\description{
Returns a mean for a logit normal such that the mode will be centered around mu
}
\keyword{MCMC}
\keyword{optimization}
