% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_funs.R
\name{lognormal_imode}
\alias{lognormal_imode}
\title{Get inverse log-normal mode}
\usage{
lognormal_imode(mu, sd)
}
\arguments{
\item{mu}{the value around which the mode should be centered (in log space)}

\item{sd}{the standard deviation of the lognormal distribution (in log space)}
}
\value{
the proposed mean for the distribution
}
\description{
Returns a mean for a lognormal such that the mode will be centered around mu
}
\keyword{MCMC}
\keyword{optimization}
