% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smapping_functions.R
\name{makeblock}
\alias{makeblock}
\title{Make an embedding block from timeseries data}
\usage{
makeblock(Y, E, lib = NULL)
}
\arguments{
\item{Y}{a timeseries vector from which to build the embedding.}

\item{E}{a positive integer, specifying the embedding dimension}

\item{lib}{an optional matrix of library positions, for specifying cases where
Y is a composite timeseries made up of multiple separate observations (e.g. spatial replicates).
Matrix should have two columns, with the first row in each column specifying the start of the
timeseries section, and the second column specifying the end.}
}
\value{
a matrix of time-delayed embeddings
}
\description{
Returns a matrix X, where columns are time-delayed embeddings of Y, with number of embeddings
specified by embedding dimension E.
See help file for the S_map_Sugihara1994 function for examples.
}
