% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{makedynamics}
\alias{makedynamics}
\title{Simulate fake data}
\usage{
makedynamics(n = 1000, n0 = NULL, obs = c(log(0.01), log(0.1)),
  proc = c(-2, log(1.2)), pcol = c(logit(0.2), log(0.01)),
  r = log(3), K = log(1), doplot = FALSE)
}
\arguments{
\item{n}{number of timesteps to simulate}

\item{n0}{starting population size - defaults to NULL, which starts at K/10}

\item{obs}{a numeric vector of length two (b0, b1) specifying observation error, parameterizing the function sd(obs) = b0 + b1*X, where X is the system state}

\item{proc}{a numeric vector of length two (lc, z) specifying process noise, parameterizing the function var(proc) = exp(lc + z*log(X)), where X is the system state}

\item{pcol}{a numeric vector of length two, including probability of colonization per timestep given that abundance = 0, and abundance after colonization}

\item{r}{intrinsic growth rate for May's logistic function}

\item{K}{carrying capacity for May's logistic function}

\item{doplot}{a logical specifying wether output should be plotted - defaults to FALSE}
}
\value{
An n-by-3 dataframe of states, including obs (observed values), truth (true values), and noproc (values without process noise)
}
\description{
Simulates a time series with known process noise and observation error functions. Implemented May's chaotic logistic function, X(t+1) = X(t) exp(r (1-X(t)/K)).
}
\examples{
 #run function
 datout<-makedynamics(n=2e4, proc = c(-2,1.2))

 #show regression of variance vs. mean for binned data
 datout_ps<-datout[datout$true>0 & datout$noproc>0,]
 #bins
 sq<-seq(0, quantile(datout$true, 0.95), by=0.01)
 ctd<-cut(datout_ps$noproc, sq)
 #calculate mean and variance by bin
 tdat<-data.frame(mu=(sq[-1]+sq[-length(sq)])/2,
      var=tapply((datout_ps$true-datout_ps$noproc)^2, ctd, mean),
     unk=tapply((datout_ps$true-datout_ps$noproc)^2, ctd, sd))
 #plot result
 plot(log(tdat$mu), log(tdat$var), xlab="mu", ylab="var")
 #show regression
 summary(mod<-lm(log(var)~log(mu), tdat, weights = 1/tdat$unk)); abline(mod, col=2)
}
\keyword{Taylor}
\keyword{data}
\keyword{fake}
\keyword{law,}
\keyword{power}
\keyword{stability,}
\keyword{time-series,}
