% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{obsfun0}
\alias{obsfun0}
\title{default observation noise function}
\usage{
obsfun0(so, yt, xt = NULL, inverse = FALSE, N = NULL)
}
\arguments{
\item{so}{a numeric vector of length two (a, b), specifying intercept and slope for linear observation error function}

\item{yt}{a number, representing a potential observed value of xt}

\item{xt}{a number or numeric vector of "true" (or simulated) abundances at time t, from which the likelihood of yt will be calculated - defaults to NULL for inverse=TRUE}

\item{inverse}{a logical specifying whether inverse (i.e. random number generator) function should be implemented - defaults to FALSE}

\item{N}{number of draws from the random number generator, if inverse=TRUE - defaults to NULL}
}
\value{
If inverse=FALSE, a number or numeric vector of length xt, with predicted log likelihoods of observation yt. If inverse = FALSE, returns N random draws from the observation function.
}
\description{
Two options: If inverse=FALSE, calculates the log probability density of observation yt based on true state xt and observation error.
Otherwise, simulates N random observations of yt.
Observation error follows a linear function of abundance, such that sd(xt) = exp(a) + exp(b)*xt.
Note that probability density is calculated based on a Tobit distribution, with lower boundary zero.
}
\keyword{error}
\keyword{observation}
