% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{parseparam0}
\alias{parseparam0}
\title{Parse parameters}
\usage{
parseparam0(param, detparam = c(log(3), log(1)))
}
\arguments{
\item{param}{List of 6 or 8 paramters, in the order (obs1, obs2, proc1, proc2, pcol1, pcol2, det1, det2).
Note that if param is of length 6, then detparam must be supplied. See obsfun0, procfun0, and detfun0 for more details.}

\item{detparam}{Optional vector of length two, including paramters for the deterministic function.}
}
\value{
a formatted list of parameters
}
\description{
Takes in a vector of 6 or 8 parameters, and puts them into a list of the format expected by the particleFilterLL function.
}
\keyword{filter}
\keyword{particle}
\keyword{stability,}
\keyword{time-series,}
