% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{particleFilterLL_piecewise}
\alias{particleFilterLL_piecewise}
\title{run particle filter across piecewise data}
\usage{
particleFilterLL_piecewise(param, N, y, libuse_y, smap_coefs, Euse, tuse,
  colpar = c(logit(1e-06), log(0.1)), nsmp = 1)
}
\arguments{
\item{param}{parameters to be passed to parseparam0 function}

\item{N}{number of particles}

\item{y}{the time series to be analyzed}

\item{libuse_y}{a matrix with two columns, specifying the start end end positions of segments within vector y}

\item{smap_coefs}{a matrix of s-mapping coefficients}

\item{Euse}{embedding dimension for the s-mapping analysis}

\item{tuse}{theta for s-mapping analysis}

\item{colpar}{parameters to be passed to the colfun0 - defaults to c(logit(1e-6), log(0.1))}

\item{nsmp}{number of sample particle trajectories to return - defaults to 1}
}
\value{
results from particle filter - including mean estimates (Nest) and standard deviations (Nsd), across particles,
and sample particle trajectories with (Nsmp) and without (Nsmp_noproc) process noise
}
\description{
Calculates likelihoods across several segments of data - e.g. multiple plots from a single experiment.
Requires several implicitely defined variables to run:
}
\keyword{dewdrop}
\keyword{filter}
\keyword{particle}
\keyword{regression,}
