% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{process_scof}
\alias{process_scof}
\title{Process s-mapping coefficients}
\usage{
process_scof(smap_coefs)
}
\arguments{
\item{smap_coefs}{a matrix of s-map coefficients, taken from the s_map function.}
}
\value{
a matrix of s-mapping coefficients
}
\description{
Processes s-mapping coefficients from rEDM into a matrix of form C1, C2, C3, ... C0, where C0 is the intercept,
C1 is the current time step t, C2 is timestep t-1, C3 is timestep t-2, and so on.
Rows correspond to the time step used to produce the prediction, e.g. row 4 is used to calculate
predicted value for time step 5. This is the format expected by the EDMfun0 function.
Note - the format produced by the rEDM package has changed substantially over time,
and so if you find that predictions are no longer working in this package, it is likely that this is
due to another change in reporting format, in which case you may need to update this function
accordingly (e.g. to re-align columns or rows).
}
