% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{procfun0}
\alias{procfun0}
\title{default process noise function}
\usage{
procfun0(sp, xt)
}
\arguments{
\item{sp}{a numeric vector of length two (a, z), specifying intercept and slope for Taylor Power scaling relationship}

\item{xt}{a number or numeric vector of abundances at time t, before process noise has occurred}
}
\value{
a number or numeric vector of length xt, with predicted abundances after process noise has occurred
}
\description{
Simulates effects of process noise following the Taylor Power Law, i.e. with standard devation sd(xt) = sqrt(exp(exp(a)+exp(z)*log(xt))).
Note that process noise only influences positive abundances (i.e. process noise cannot contribute to colonization)
}
\keyword{Taylor}
\keyword{law}
\keyword{noise,}
\keyword{power}
\keyword{process}
