% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{sampler_fun0}
\alias{sampler_fun0}
\title{Default sampler function for prior}
\usage{
sampler_fun0(n = 1, pars = pars, priorsd = c(0.5, 0.5, 2, 0.5, 0.5,
  0.5))
}
\arguments{
\item{n}{number of random draws to take from the priors}

\item{pars}{a list of parameter values, structured following the parseparam0 function, specifying the centers for the priors}

\item{priorsd}{a vector of length 6 or 8 with desired standard deviations for the prior distributions}
}
\value{
returns random draws from the priors
}
\description{
Text...
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability,}
\keyword{time-series,}
