% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{sdproc_abstract}
\alias{sdproc_abstract}
\title{calculate estimated total variance}
\usage{
sdproc_abstract(sd_proc, rgr, waiting_time = 1)
}
\arguments{
\item{sd_proc}{standard deviation of the (Gaussian) disturbance process}

\item{rgr}{relative growth rate of the linear process}

\item{waiting_time}{average waiting time between (random exponentially distributed through time) disturbance events}
}
\value{
standard deviation of stochastic variability in x
}
\description{
Function for estimating stochastic variation in linar process x as a function of relative growth rate and disturbance regime standard deviation.
}
\keyword{linear}
\keyword{stochastic}
\keyword{system}
