% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{density_fun0}
\alias{density_fun0}
\title{Default density function for prior}
\usage{
density_fun0(param, minv, maxv)
}
\arguments{
\item{param}{a vector model parameters}

\item{minv}{a vector of minimum values for the interval}

\item{maxv}{a vector of maximum values for the interval}
}
\value{
returns log likelihood of parameters given priors.
}
\description{
Default density function, following the syntax for priors in the BayesianTools package. Uses
flat priors for all paramters, within the given interval. Density function integrates to one.
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability}
\keyword{time-series}
