% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{inv_fun0}
\alias{inv_fun0}
\title{Default inverse transormation function}
\usage{
inv_fun0(x)
}
\arguments{
\item{x}{an nxm matrix with}
}
\value{
returns back-transformed values of parameters
}
\description{
Takes in a matrix, where each column represents a parameter. Returns parameters in untransformed space.
If length = 2, then in the order (obs1, proc1).
If 3, then in the order (obs1, proc1, proc2).
If 4, then in the order (obs1, obs2, proc1, proc2).
If 6, then in the order (obs1, proc1, pcol1, pcol2, det1, det2)
If 7, then in the order (obs1, proc1, proc2, pcol1, pcol2, det1, det2)
If 8, then in the order (obs1, obs2, proc1, proc2, pcol1, pcol2, det1, det2)
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability}
\keyword{time-series}
