% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{parseparam0}
\alias{parseparam0}
\title{Parse parameters}
\usage{
parseparam0(
  param,
  colparam = c(logit(0.2), log(0.1)),
  detparam = c(log(1.2), log(1))
)
}
\arguments{
\item{param}{List of paramters, of length 2, 3, 4, 6, 7, or 8.
If 2, then in the order (obs1, proc1).
If 3, then in the order (obs1, proc1, proc2).
If 4, then in the order (obs1, obs2, proc1, proc2).
If 6, then in the order (obs1, proc1, pcol1, pcol2, det1, det2)
If 7, then in the order (obs1, proc1, proc2, pcol1, pcol2, det1, det2)
If 8, then in the order (obs1, obs2, proc1, proc2, pcol1, pcol2, det1, det2)
Note that if param is of length 2 or 3, then detparam  and colparam must be supplied. See obsfun0, procfun0, and detfun0 for more details.}

\item{colparam}{Optional vector of length two, including parameters for the colonization function.}

\item{detparam}{Optional vector of length two, including paramters for the deterministic function.}
}
\value{
a formatted list of parameters
}
\description{
Takes in a vector of 3 or 6 parameters, and puts them into a list of the format expected by the particleFilterLL function.
}
\keyword{filter}
\keyword{particle}
\keyword{stability}
\keyword{time-series}
