% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smapping_functions.R
\name{process_scof}
\alias{process_scof}
\title{Process S-mapping coefficients}
\usage{
process_scof(smap_coefs)
}
\arguments{
\item{smap_coefs}{a matrix of s-map coefficients, taken from the S_map_Sugihara1994 function.}
}
\value{
a matrix of s-mapping coefficients
}
\description{
Processes s-mapping coefficients from S_map_Sugihara1994 into a matrix of form C1, C2, C3, ... C0, where C0 is the intercept,
C1 is the current time step t, C2 is timestep t-1, C3 is timestep t-2, and so on.
Rows correspond to the time step used to produce the prediction, e.g. row 4 is used to calculate
predicted value for time step 5. This is the format expected by the EDMfun0 function.
See help file for the S_map_Sugihara1994 function for examples.
}
