% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfun.R
\name{sampler_fun0}
\alias{sampler_fun0}
\title{Default sampler function for prior}
\usage{
sampler_fun0(n = 1, minv, maxv)
}
\arguments{
\item{n}{number of random draws to take from the priors}

\item{minv}{Vector of minimum values to return for each parameter}

\item{maxv}{Vector of maximum values to return for each parameter}
}
\value{
returns random draws from the priors
}
\description{
Draws samples from a flat prior
}
\keyword{MCMC}
\keyword{optimization}
\keyword{stability}
\keyword{time-series}
